requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
openai==1.13.3
google-generativeai==0.3.0
